+++ src/main/java/com/openbook/openbook/booth/controller/BoothProductController.java
@@ -1,0 +1,62 @@
+package com.openbook.openbook.booth.controller;
+
+import com.openbook.openbook.booth.controller.request.ProductCategoryRegister;
+import com.openbook.openbook.booth.controller.request.ProductRegistrationRequest;
+import com.openbook.openbook.booth.controller.response.CategoryProductsResponse;
+import com.openbook.openbook.booth.controller.response.ProductCategoryResponse;
+import com.openbook.openbook.booth.service.BoothProductService;
+import com.openbook.openbook.global.dto.ResponseMessage;
+import jakarta.validation.Valid;
+import java.util.List;
+import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.web.PageableDefault;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequiredArgsConstructor
+public class BoothProductController {
+
+    private final BoothProductService boothProductService;
+
+    @PostMapping("/booths/{booth_id}/product-category")
+    public ResponseEntity<ResponseMessage> addProductCategory(Authentication authentication,
+                                                              @PathVariable Long booth_id,
+                                                              @Valid ProductCategoryRegister request) {
+        boothProductService.addProductCategory(Long.valueOf(authentication.getName()), booth_id, request);
+        return ResponseEntity.ok(new ResponseMessage("상품 카테고리 생성에 성공했습니다."));
+    }
+
+    @PostMapping("/booths/{booth_id}/products")
+    public ResponseEntity<ResponseMessage> addProduct(Authentication authentication,
+                                                      @PathVariable Long booth_id,
+                                                      @Valid ProductRegistrationRequest request){
+        boothProductService.addBoothProduct(Long.valueOf(authentication.getName()), booth_id, request);
+        return ResponseEntity.status(HttpStatus.CREATED).body(new ResponseMessage("상품 추가에 성공했습니다."));
+    }
+
+    @GetMapping("/{booth_id}/product-category")
+    public ResponseEntity<List<ProductCategoryResponse>> getProductCategory(@PathVariable Long booth_id) {
+        return ResponseEntity.ok(boothProductService.getProductCategoryResponseList(booth_id));
+    }
+
+    @GetMapping("/{booth_id}/products")
+    public ResponseEntity<List<CategoryProductsResponse>> getAllBoothProducts(@PathVariable Long booth_id,
+                                                                              @PageableDefault(size = 5) Pageable pageable) {
+        return ResponseEntity.ok(boothProductService.findAllBoothProducts(booth_id, pageable));
+    }
+
+    @GetMapping("/products/category")
+    public ResponseEntity<CategoryProductsResponse> getProductsByCategory(@RequestParam Long category_id,
+                                                                          @PageableDefault(size = 5) Pageable pageable) {
+        return ResponseEntity.ok(boothProductService.findCategoryProducts(category_id, pageable));
+    }
+
+}
